restructure settings modal tabs and fields
reorder tabs to profile > account > privacy, merge password
into account, make username/email read-only with change button,
move display name to profile, move element/language/theme to
account, replace dividers with section headers, widen settings
row control area
