+++ messages/en.json
@@ -1003,6 +1003,7 @@
 	"settings_username": "Username",
 	"settings_username_placeholder": "Enter username",
 	"settings_display_name": "Display name",
+	"settings_display_name_subtitle": "The name that appears on your profile",
 	"settings_display_name_placeholder": "Enter display name",
 	"settings_email": "Email",
 	"settings_email_placeholder": "Enter email",
@@ -1037,7 +1038,9 @@
 	"settings_youtube": "YouTube",
 	"settings_youtube_subtitle": "Your YouTube channel handle",
 	"settings_youtube_placeholder": "Enter handle",
-	"settings_gender": "Gender",
+	"settings_section_teams": "Teams",
+	"settings_section_elsewhere": "Elsewhere",
+	"settings_gender": "MC Gender",
 	"settings_gender_subtitle": "Your in-game character",
 	"settings_gender_placeholder": "Select gender",
 	"settings_language": "Language",
