+++ messages/ja.json
@@ -1003,6 +1003,7 @@
 	"settings_username": "ユーザー名",
 	"settings_username_placeholder": "ユーザー名を入力",
 	"settings_display_name": "表示名",
+	"settings_display_name_subtitle": "プロフィールに表示される名前",
 	"settings_display_name_placeholder": "表示名を入力",
 	"settings_email": "メールアドレス",
 	"settings_email_placeholder": "メールアドレスを入力",
@@ -1037,6 +1038,8 @@
 	"settings_youtube": "YouTube",
 	"settings_youtube_subtitle": "YouTubeチャンネルのハンドル",
 	"settings_youtube_placeholder": "ハンドルを入力",
+	"settings_section_teams": "チーム",
+	"settings_section_elsewhere": "その他",
 	"settings_gender": "主人公",
 	"settings_gender_subtitle": "ゲーム内の主人公キャラクター",
 	"settings_gender_placeholder": "主人公を選択",
