+++ src/lib/components/settings/AccountSettings.svelte
@@ -2,40 +2,66 @@
 <script lang="ts">
 	import * as m from '$lib/paraglide/messages'
 	import Input from '../ui/Input.svelte'
+	import Button from '../ui/Button.svelte'
+	import Select from '../ui/Select.svelte'
 	import Switch from '../ui/switch/Switch.svelte'
 	import SettingsRow from '../ui/SettingsRow.svelte'
+	import ElementPicker from '../ui/element-picker/ElementPicker.svelte'
 	import type { ElementType } from '../ui/SettingsNav.svelte'
+	import { getElementKey } from '$lib/utils/element'
 	import { userAdapter } from '$lib/api/adapters/user.adapter'
 
 	interface Props {
 		username: string
-		displayName: string
 		email: string
 		emailVerified: boolean
 		bahamut: boolean
 		role: number
 		element: ElementType
+		language: string
+		theme: string
+		currentPassword: string
+		newPassword: string
+		confirmPassword: string
 		onUsernameChange: (value: string) => void
-		onDisplayNameChange: (value: string) => void
 		onEmailChange: (value: string) => void
 		onBahamutChange: (value: boolean) => void
 		onUsernameValidChange?: (valid: boolean) => void
+		onElementChange: (value: string) => void
+		onLanguageChange: (value: string) => void
+		onThemeChange: (value: string) => void
+		onCurrentPasswordChange: (value: string) => void
+		onNewPasswordChange: (value: string) => void
+		onConfirmPasswordChange: (value: string) => void
 	}
 
 	let {
 		username,
-		displayName,
 		email,
 		emailVerified,
 		bahamut,
 		role,
 		element,
+		language,
+		theme,
+		currentPassword,
+		newPassword,
+		confirmPassword,
 		onUsernameChange,
-		onDisplayNameChange,
 		onEmailChange,
 		onBahamutChange,
-		onUsernameValidChange
+		onUsernameValidChange,
+		onElementChange,
+		onLanguageChange,
+		onThemeChange,
+		onCurrentPasswordChange,
+		onNewPasswordChange,
+		onConfirmPasswordChange
 	}: Props = $props()
+
+	// Editing state for read-only fields
+	let editingUsername = $state(false)
+	let editingEmail = $state(false)
 
 	let resending = $state(false)
 	let resendMessage = $state<string | null>(null)
@@ -55,7 +81,6 @@
 
 	// Local state initialized from props — mutable for form editing
 	let localUsername = $state(username)
-	let localDisplayName = $state(displayName)
 	let localEmail = $state(email)
 
 	// Capture original username for comparison (prop gets mutated by parent on each keystroke)
@@ -144,76 +169,175 @@
 			usernameTimer = setTimeout(() => checkUsernameAvailability(localUsername), 300)
 		}
 	}
-	function handleDisplayNameInput() {
-		onDisplayNameChange(localDisplayName)
-	}
 	function handleEmailInput() {
 		onEmailChange(localEmail)
 	}
 
 	// Check if user is admin
 	const isAdmin = $derived(role === 9)
+
+	// Element key ↔ numeric ID conversion
+	const ELEMENT_KEY_TO_ID: Record<string, number> = {
+		wind: 1, fire: 2, water: 3, earth: 4, dark: 5, light: 6
+	}
+	const elementId = $derived(ELEMENT_KEY_TO_ID[element] ?? 1)
+
+	// Language/Theme local state
+	let localLanguage = $derived(language)
+	let localTheme = $derived(theme)
+
+	$effect(() => {
+		if (localLanguage !== language) onLanguageChange(localLanguage)
+	})
+	$effect(() => {
+		if (localTheme !== theme) onThemeChange(localTheme)
+	})
+
+	const languageOptions = [
+		{ value: 'en', label: 'English' },
+		{ value: 'ja', label: '日本語' }
+	]
+
+	const themeOptions = [
+		{ value: 'system', label: m.settings_theme_system() },
+		{ value: 'light', label: m.settings_theme_light() },
+		{ value: 'dark', label: m.settings_theme_dark() }
+	]
+
+	// Password local state
+	let localCurrentPassword = $state(currentPassword)
+	let localNewPassword = $state(newPassword)
+	let localConfirmPassword = $state(confirmPassword)
+
+	function handleCurrentPasswordInput() {
+		onCurrentPasswordChange(localCurrentPassword)
+	}
+	function handleNewPasswordInput() {
+		onNewPasswordChange(localNewPassword)
+	}
+	function handleConfirmPasswordInput() {
+		onConfirmPasswordChange(localConfirmPassword)
+	}
+
+	const hasSecurityChanges = $derived(localNewPassword !== '' || localConfirmPassword !== '')
+	const passwordsMatch = $derived(localNewPassword === '' || localNewPassword === localConfirmPassword)
+	const passwordError = $derived(!passwordsMatch ? m.settings_password_mismatch() : '')
+	const currentPasswordRequired = $derived(hasSecurityChanges && localCurrentPassword === '')
 </script>
 
 <div class="section">
 	<div class="form-fields">
 		<!-- Username -->
-		<Input
-			label={m.settings_username()}
-			placeholder={m.settings_username_placeholder()}
-			contained
-			fullWidth
-			bind:value={localUsername}
-			handleInput={handleUsernameInput}
-			error={usernameError}
-			rightIcon={usernameIcon}
-		/>
-
-		<!-- Display Name -->
-		<Input
-			label={m.settings_display_name()}
-			placeholder={m.settings_display_name_placeholder()}
-			contained
-			fullWidth
-			bind:value={localDisplayName}
-			handleInput={handleDisplayNameInput}
-		/>
-
-		<!-- Email -->
-		<div class="email-group">
+		{#if editingUsername}
 			<Input
-				label={m.settings_email()}
-				type="email"
-				placeholder={m.settings_email_placeholder()}
+				label={m.settings_username()}
+				placeholder={m.settings_username_placeholder()}
 				contained
 				fullWidth
-				bind:value={localEmail}
-				handleInput={handleEmailInput}
+				bind:value={localUsername}
+				handleInput={handleUsernameInput}
+				error={usernameError}
+				rightIcon={usernameIcon}
 			/>
-
-			<!-- Email verification status -->
-			<div class="verification-status">
-			{#if emailVerified}
-				<span class="verified">{m.settings_email_verified()}</span>
-			{:else}
-				<span class="unverified">{m.settings_email_unverified()}</span>
-				<button
-					class="resend-link"
-					onclick={handleResendVerification}
-					disabled={resending}
-				>
-					{resending ? m.settings_verification_resending() : m.settings_verification_resend()}
-				</button>
-				{#if resendMessage}
-					<span class="resend-message">{resendMessage}</span>
-				{/if}
-			{/if}
+		{:else}
+			<div class="readonly-row">
+				<div class="readonly-field">
+					<span class="field-label">{m.settings_username()}</span>
+					<span class="field-value">{username}</span>
+				</div>
+				<Button variant="text" size="small" onclick={() => (editingUsername = true)}>
+					{m.action_change()}
+				</Button>
 			</div>
-		</div>
+		{/if}
+
+		<!-- Email -->
+		{#if editingEmail}
+			<div class="email-group">
+				<Input
+					label={m.settings_email()}
+					type="email"
+					placeholder={m.settings_email_placeholder()}
+					contained
+					fullWidth
+					bind:value={localEmail}
+					handleInput={handleEmailInput}
+				/>
+
+				<!-- Email verification status -->
+				<div class="verification-status">
+				{#if emailVerified}
+					<span class="verified">{m.settings_email_verified()}</span>
+				{:else}
+					<span class="unverified">{m.settings_email_unverified()}</span>
+					<button
+						class="resend-link"
+						onclick={handleResendVerification}
+						disabled={resending}
+					>
+						{resending ? m.settings_verification_resending() : m.settings_verification_resend()}
+					</button>
+					{#if resendMessage}
+						<span class="resend-message">{resendMessage}</span>
+					{/if}
+				{/if}
+				</div>
+			</div>
+		{:else}
+			<div class="readonly-row">
+				<div class="readonly-field">
+					<span class="field-label">{m.settings_email()}</span>
+					<span class="field-value">{email}</span>
+				</div>
+				<Button variant="text" size="small" onclick={() => (editingEmail = true)}>
+					{m.action_change()}
+				</Button>
+			</div>
+		{/if}
+
+		<!-- Element Selection -->
+		<SettingsRow title={m.settings_element()} subtitle={m.settings_element_subtitle()}>
+			{#snippet control()}
+				<ElementPicker
+					value={elementId}
+					onValueChange={(v) => {
+						const key = getElementKey(v as number)
+						onElementChange(key)
+					}}
+					mode="dropdown"
+					contained
+				/>
+			{/snippet}
+		</SettingsRow>
+
+		<!-- Language Selection -->
+		<SettingsRow title={m.settings_language()} subtitle={m.settings_language_subtitle()}>
+			{#snippet control()}
+				<Select
+					bind:value={localLanguage}
+					options={languageOptions}
+					placeholder={m.settings_language_placeholder()}
+					contained
+					portal
+				/>
+			{/snippet}
+		</SettingsRow>
+
+		<!-- Theme Selection -->
+		<SettingsRow title={m.settings_theme()} subtitle={m.settings_theme_subtitle()}>
+			{#snippet control()}
+				<Select
+					bind:value={localTheme}
+					options={themeOptions}
+					placeholder={m.settings_theme_placeholder()}
+					contained
+					portal
+				/>
+			{/snippet}
+		</SettingsRow>
 
 		<!-- Bahamut Mode (admin only) -->
 		{#if isAdmin}
-			<hr class="separator" />
 			<SettingsRow title={m.settings_bahamut_mode()} subtitle={m.settings_bahamut_subtitle()}>
 				{#snippet control()}
 					<Switch
@@ -225,6 +349,46 @@
 				{/snippet}
 			</SettingsRow>
 		{/if}
+
+		<!-- Password Section -->
+		<h3 class="section-header">{m.settings_nav_password()}</h3>
+
+		<p class="section-note">
+			{m.settings_password_note()}
+		</p>
+
+		<Input
+			label={m.settings_current_password()}
+			type="password"
+			placeholder={m.settings_current_password_placeholder()}
+			contained
+			fullWidth
+			required={hasSecurityChanges}
+			error={currentPasswordRequired ? m.settings_current_password_required() : ''}
+			bind:value={localCurrentPassword}
+			handleInput={handleCurrentPasswordInput}
+		/>
+
+		<Input
+			label={m.settings_new_password()}
+			type="password"
+			placeholder={m.settings_new_password_placeholder()}
+			contained
+			fullWidth
+			bind:value={localNewPassword}
+			handleInput={handleNewPasswordInput}
+		/>
+
+		<Input
+			label={m.settings_confirm_password()}
+			type="password"
+			placeholder={m.settings_confirm_password_placeholder()}
+			contained
+			fullWidth
+			error={passwordError}
+			bind:value={localConfirmPassword}
+			handleInput={handleConfirmPasswordInput}
+		/>
 	</div>
 </div>
 
@@ -243,9 +407,39 @@
 		gap: spacing.$unit-3x;
 	}
 
-	.separator {
-		border: none;
-		border-top: 1px solid var(--separator-bg);
+	.section-header {
+		font-size: typography.$font-small;
+		font-weight: typography.$medium;
+		color: var(--text-secondary);
+		margin: 0;
+	}
+
+	.readonly-row {
+		display: flex;
+		align-items: center;
+		justify-content: space-between;
+		gap: spacing.$unit-2x;
+	}
+
+	.readonly-field {
+		display: flex;
+		flex-direction: column;
+		gap: spacing.$unit-half;
+	}
+
+	.field-label {
+		font-size: typography.$font-small;
+		color: var(--text-secondary);
+	}
+
+	.field-value {
+		font-size: typography.$font-regular;
+		color: var(--text-primary);
+	}
+
+	.section-note {
+		font-size: typography.$font-small;
+		color: var(--text-secondary);
 		margin: 0;
 	}
