+++ src/lib/components/settings/ProfileSettings.svelte
@@ -4,56 +4,44 @@
 	import Select from '../ui/Select.svelte'
 	import Input from '../ui/Input.svelte'
 	import SettingsRow from '../ui/SettingsRow.svelte'
-	import ElementPicker from '../ui/element-picker/ElementPicker.svelte'
 	import { pictureData } from '$lib/utils/pictureData'
 	import { localizedName } from '$lib/utils/locale'
 	import { getAvatarSrc, getAvatarSrcSet } from '$lib/utils/avatar'
-	import { getElementKey } from '$lib/utils/element'
 	import type { ElementType } from '../ui/SettingsNav.svelte'
 
 	interface Props {
 		picture: string
 		element: ElementType
+		username: string
+		displayName: string
 		granblueId: string
 		wikiProfile: string
 		youtube: string
 		gender: number
-		language: string
-		theme: string
 		onPictureChange: (value: string) => void
-		onElementChange: (value: string) => void
+		onDisplayNameChange: (value: string) => void
 		onGranblueIdChange: (value: string) => void
 		onWikiProfileChange: (value: string) => void
 		onYoutubeChange: (value: string) => void
 		onGenderChange: (value: number) => void
-		onLanguageChange: (value: string) => void
-		onThemeChange: (value: string) => void
 	}
 
 	let {
 		picture,
 		element,
+		username,
+		displayName,
 		granblueId,
 		wikiProfile,
 		youtube,
 		gender,
-		language,
-		theme,
 		onPictureChange,
-		onElementChange,
+		onDisplayNameChange,
 		onGranblueIdChange,
 		onWikiProfileChange,
 		onYoutubeChange,
-		onGenderChange,
-		onLanguageChange,
-		onThemeChange
+		onGenderChange
 	}: Props = $props()
-
-	// Element key ↔ numeric ID conversion
-	const ELEMENT_KEY_TO_ID: Record<string, number> = {
-		wind: 1, fire: 2, water: 3, earth: 4, dark: 5, light: 6
-	}
-	const elementId = $derived(ELEMENT_KEY_TO_ID[element] ?? 1)
 
 	// Prepare options for selects
 	const pictureOptions = $derived(
@@ -71,36 +59,23 @@
 		{ value: 1, label: 'Djeeta' }
 	]
 
-	const languageOptions = [
-		{ value: 'en', label: 'English' },
-		{ value: 'ja', label: '日本語' }
-	]
-
-	const themeOptions = [
-		{ value: 'system', label: m.settings_theme_system() },
-		{ value: 'light', label: m.settings_theme_light() },
-		{ value: 'dark', label: m.settings_theme_dark() }
-	]
-
 	// Get current picture data
 	const currentPicture = $derived(pictureData.find((p) => p.filename === picture))
 
 	// Local state derived from props — overrides via bind:value are temporary
 	let localPicture = $derived(picture)
-	let localElement = $derived(element)
+	let localDisplayName = $derived(displayName)
 	let localGranblueId = $derived(granblueId)
 	let localWikiProfile = $derived(wikiProfile)
 	let localYoutube = $derived(youtube)
 	let localGender = $derived(gender)
-	let localLanguage = $derived(language)
-	let localTheme = $derived(theme)
 
 	// Propagate changes
 	$effect(() => {
 		if (localPicture !== picture) onPictureChange(localPicture)
 	})
 	$effect(() => {
-		if (localElement !== element) onElementChange(localElement)
+		if (localDisplayName !== displayName) onDisplayNameChange(localDisplayName)
 	})
 	$effect(() => {
 		if (localGranblueId !== granblueId) onGranblueIdChange(localGranblueId)
@@ -114,24 +89,18 @@
 	$effect(() => {
 		if (localGender !== gender) onGenderChange(localGender)
 	})
-	$effect(() => {
-		if (localLanguage !== language) onLanguageChange(localLanguage)
-	})
-	$effect(() => {
-		if (localTheme !== theme) onThemeChange(localTheme)
-	})
 </script>
 
 <div class="section">
 	<div class="form-fields">
-		<!-- Picture Selection with Preview -->
+		<!-- Avatar + Display Name (untitled section) -->
 		<div class="picture-section">
 			<div class="current-avatar">
 				<img
 					src={getAvatarSrc(localPicture)}
 					srcset={getAvatarSrcSet(localPicture)}
 					alt={currentPicture ? localizedName(currentPicture.name) : ''}
-					class="avatar-preview element-{localElement}"
+					class="avatar-preview element-{element}"
 				/>
 			</div>
 			<Select
@@ -145,47 +114,15 @@
 			/>
 		</div>
 
-		<!-- Element Selection -->
-		<SettingsRow title={m.settings_element()} subtitle={m.settings_element_subtitle()}>
+		<SettingsRow title={m.settings_display_name()} subtitle={m.settings_display_name_subtitle()}>
 			{#snippet control()}
-				<ElementPicker
-					value={elementId}
-					onValueChange={(v) => {
-						const key = getElementKey(v as number)
-						localElement = key as ElementType
-					}}
-					mode="dropdown"
-					contained
-				/>
+				<Input bind:value={localDisplayName} placeholder={displayName ? m.settings_display_name_placeholder() : username} contained />
 			{/snippet}
 		</SettingsRow>
 
-		<hr class="separator" />
+		<!-- Teams -->
+		<h3 class="section-header">{m.settings_section_teams()}</h3>
 
-		<!-- Granblue ID -->
-		<SettingsRow title={m.settings_granblue_id()} subtitle={m.settings_granblue_id_subtitle()}>
-			{#snippet control()}
-				<Input bind:value={localGranblueId} placeholder={m.settings_granblue_id_placeholder()} contained />
-			{/snippet}
-		</SettingsRow>
-
-		<!-- Wiki Profile -->
-		<SettingsRow title={m.settings_wiki_profile()} subtitle={m.settings_wiki_profile_subtitle()}>
-			{#snippet control()}
-				<Input bind:value={localWikiProfile} placeholder={m.settings_wiki_profile_placeholder()} contained />
-			{/snippet}
-		</SettingsRow>
-
-		<!-- YouTube -->
-		<SettingsRow title={m.settings_youtube()} subtitle={m.settings_youtube_subtitle()}>
-			{#snippet control()}
-				<Input bind:value={localYoutube} placeholder={m.settings_youtube_placeholder()} contained />
-			{/snippet}
-		</SettingsRow>
-
-		<hr class="separator" />
-
-		<!-- Gender Selection -->
 		<SettingsRow title={m.settings_gender()} subtitle={m.settings_gender_subtitle()}>
 			{#snippet control()}
 				<Select
@@ -198,29 +135,24 @@
 			{/snippet}
 		</SettingsRow>
 
-		<!-- Language Selection -->
-		<SettingsRow title={m.settings_language()} subtitle={m.settings_language_subtitle()}>
+		<!-- Elsewhere -->
+		<h3 class="section-header">{m.settings_section_elsewhere()}</h3>
+
+		<SettingsRow title={m.settings_granblue_id()} subtitle={m.settings_granblue_id_subtitle()}>
 			{#snippet control()}
-				<Select
-					bind:value={localLanguage}
-					options={languageOptions}
-					placeholder={m.settings_language_placeholder()}
-					contained
-					portal
-				/>
+				<Input bind:value={localGranblueId} placeholder={m.settings_granblue_id_placeholder()} contained />
 			{/snippet}
 		</SettingsRow>
 
-		<!-- Theme Selection -->
-		<SettingsRow title={m.settings_theme()} subtitle={m.settings_theme_subtitle()}>
+		<SettingsRow title={m.settings_wiki_profile()} subtitle={m.settings_wiki_profile_subtitle()}>
 			{#snippet control()}
-				<Select
-					bind:value={localTheme}
-					options={themeOptions}
-					placeholder={m.settings_theme_placeholder()}
-					contained
-					portal
-				/>
+				<Input bind:value={localWikiProfile} placeholder={m.settings_wiki_profile_placeholder()} contained />
+			{/snippet}
+		</SettingsRow>
+
+		<SettingsRow title={m.settings_youtube()} subtitle={m.settings_youtube_subtitle()}>
+			{#snippet control()}
+				<Input bind:value={localYoutube} placeholder={m.settings_youtube_placeholder()} contained />
 			{/snippet}
 		</SettingsRow>
 	</div>
@@ -242,9 +174,10 @@
 		gap: spacing.$unit-3x;
 	}
 
-	.separator {
-		border: none;
-		border-top: 1px solid var(--separator-bg);
+	.section-header {
+		font-size: typography.$font-small;
+		font-weight: typography.$medium;
+		color: var(--text-secondary);
 		margin: 0;
 	}
